+++ tests/helpers/schemas/site-tree-schema.js
